Update README.md
Introduction

In this project, you are going to develop a library manager program. You need to write a server and a client program, in which the client and the server communicate through a TCP socket. Carefully go through the requirements described below. Each of the requirements (e.g. R1, R2, etc.) will be evaluated.

2. Requirements

R1. The server program (Server.java) makes use of a file named “books.txt”. It is a text file that records all the books in the library. Each line in the books.txt file is a book entry that consists of title, author, and ID of the user who is currently borrowing the book (borrower). The fields are separated by tab (‘\t’). We assume that the tab character does not appear in the title, author, or borrower. The books are sorted in case-insensitive alphabetical order of the titles. An example books.txt is as follows.